+++ core/src/main/scala/org/apache/spark/deploy/master/MasterRestAuthMode.scala
@@ -1,0 +1,40 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.spark.deploy.master
+
+private[master] object MasterRestAuthMode extends Enumeration {
+    val NoneOption = "None"
+    val SecureGatewayOption = "SecureGateway"
+
+    def serverSecuredByAlternativeMethod(mode: MasterRestAuthMode): Boolean = {
+      mode match {
+        case SecureGatewayOption => true,
+        case _ => false
+      }
+    }
+
+    def fromStringOrNone(mode: String): MasterRestAuthMode = {
+      mode.toLowerCase match {
+        case MasterRestAuthMode.toLowerCase => SecureGatewayOption
+        case NoneOption.toLowerCase => NoneOption
+        case _ =>
+          logWarning(log"Specified master rest auth mode $mode was unrecognised. Defaulting to None.")
+          None
+      }
+    }
+  }
