+++ core/src/main/scala/org/apache/spark/internal/config/package.scala
@@ -1936,9 +1936,9 @@ package object config {
     .createWithDefault(false)
 
   private[spark] val MASTER_REST_SERVER_AUTH_MODE = ConfigBuilder("spark.master.rest.auth.mode")
-    .doc("Specifies the authentication mechanism of the master REST services. The default value is None. " + 
+    .doc("Specifies the authentication mechanism of the master REST services. The default value is \"None\". " + 
       "The value \"SecureGateway\" can be used to signify that you have provided an external mechanism of " + 
-      "providing authentication for all REST API's of the spark master servers(s). This setting relates to " + 
+      "providing authentication for all REST APIs of the spark master servers(s). This setting relates to " + 
       "spark.authenticate.*, but influences only the REST interfaces of the spark master.")
     .version("4.0.0")
     .stringConf
